requirements.txt
absl-py==0.10.0
albumentations==1.3.0
astunparse==1.6.3
awscli==1.25.15
blinker==1.6.2
boto3==1.24.15
botocore==1.27.15
brotlipy==0.7.0
cachetools==5.3.0
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi @ file:///C:/b/abs_49n3v2hyhr/croot/cffi_1670423218144/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.1.14
contourpy==1.0.7
cryptography @ file:///C:/b/abs_8ecplyc3n2/croot/cryptography_1677533105000/work
cycler==0.11.0
docker==5.0.3
docker-pycreds==0.4.0
docutils==0.16
easydict==1.10
filelock==3.9.0
Flask==2.3.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
flatbuffers==1.12
fonttools==4.39.3
gast==0.3.3
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.17.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.32.0
h5py==2.10.0
huggingface-hub==0.12.1
humanfriendly==10.0
idna @ file:///C:/b/abs_bdhbebrioa/croot/idna_1666125572046/work
imageio==2.26.0
imgaug==0.4.0
importlib-metadata==6.6.0
importlib-resources==5.12.0
itsdangerous==2.1.2
jax==0.4.8
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
keyring==8.7
keyrings.alt==4.1.0
kiwisolver==1.4.4
lazy_loader==0.1
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.1
matplotlib==3.7.1
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184278/work
mkl-service==2.4.0
ml-dtypes==0.1.0
mpmath==1.2.1
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.0.33
mysqlclient==2.1.1
networkx==3.0
numpy==1.22.0
oauthlib==3.2.2
onnx==1.14.0
onnxruntime==1.14.1
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
opt-einsum==3.3.0
packaging==23.0
pandas==2.0.1
pathtools==0.1.2
Pillow==8.4.0
promise==2.3
protobuf==3.20.3
psutil==5.9.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.6
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
PyMySQL==1.0.3
pyOpenSSL @ file:///C:/b/abs_552w85x1jz/croot/pyopenssl_1677607703691/work
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pyreadline3==3.4.1
PySocks @ file:///C:/ci/pysocks_1605287845585/work
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
pywifi==1.1.12
pywin32==227
pywin32-ctypes==0.2.0
PyYAML==5.4.1
qudida==0.0.4
requests @ file:///C:/b/abs_41owkd5ymz/croot/requests_1682607524657/work
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
scikit-image==0.19.3
scikit-learn==1.2.2
scipy==1.10.1
sentry-sdk==1.6.0
setproctitle==1.2.3
shapely==2.0.1
shortuuid==0.5.0
six==1.15.0
smmap==5.0.0
SQLAlchemy==2.0.13
sympy @ file:///C:/b/abs_95fbf1z7n6/croot/sympy_1668202411612/work
tabulate==0.8.10
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.4.0
tensorflow-estimator==2.4.0rc0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2023.2.28
timm==0.6.12
torch==2.0.0+cu118
torchaudio==2.0.1+cu118
torchstat==0.0.7
torchvision==0.15.1+cu118
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.25.11
voluptuous==0.13.1
wandb==0.12.19
websocket-client==1.3.3
Werkzeug==2.3.2
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306167264/work
wrapt==1.12.1
yolo==0.3.1
YOLO3==1.0
zipp==3.15.0
